+++ CHANGELOG.md
@@ -31,10 +31,10 @@ Changelog
 - Acceptance test for "Import wallet with/without spending password" feature
 - Acceptance test for "Send money from a wallet with spending password" feature
 - Acceptance test for "Generate wallet address" feature
+- Acceptance test for "Wallet settings management" features
 - Final version of Daedalus logo added on the loading screen
 - Final version of Daedalus logo added in the top-bar
 - Receive page design update
-- Acceptance test for wallet settings (Wallet password setting, Wallet setting invalid password, Wallet password update/change, Wallet password remove, Wallet rename, Wallet assurance level chage)
 
 ### Fixes